+++ whittle/training_strategies/sandwich.py
@@ -33,20 +33,27 @@ class SandwichStrategy(BaseTrainingStrategy):
     def __call__(self, model, inputs, outputs, scale_loss=1, **kwargs):
         total_loss = 0
         # update super-network
-        y_supernet = model(inputs)
-        loss = self.loss_function(y_supernet, outputs) * scale_loss
+        model.reset_super_network()
+        if self.lora:
+            loss = self.chunked_loss(model, inputs, outputs) * scale_loss
+        else:
+            y_supernet = model(inputs)
+            loss = self.loss_function(y_supernet, outputs) * scale_loss
         loss.backward() if self.fabric is None else self.fabric.backward(loss)
         total_loss += loss.item()
 
         # update random sub-networks
         for i in range(self.random_samples):
             config = self.sampler.sample()
-            model.select_sub_network(config)
+            model.set_sub_network(**config)
             y_hat = model(inputs)
             if self.kd_loss is not None:
                 loss = self.kd_loss(y_hat, outputs, y_supernet)
             else:
-                loss = self.loss_function(y_hat, outputs)
+                if self.lora:
+                    loss = self.chunked_loss(model, inputs, outputs)
+                else:
+                    loss = self.loss_function(y_hat, outputs)
 
             loss *= scale_loss
             loss.backward() if self.fabric is None else self.fabric.backward(loss)
@@ -55,12 +62,15 @@ class SandwichStrategy(BaseTrainingStrategy):
 
         # smallest network
         config = self.sampler.get_smallest_sub_network()
-        model.select_sub_network(config)
+        model.set_sub_network(**config)
         y_hat = model(inputs)
         if self.kd_loss is not None:
             loss = self.kd_loss(y_hat, outputs, y_supernet)
         else:
-            loss = self.loss_function(y_hat, outputs)
+            if self.lora:
+                loss = self.chunked_loss(model, inputs, outputs)
+            else:
+                loss = self.loss_function(y_hat, outputs)
         loss *= scale_loss
         loss.backward() if self.fabric is None else self.fabric.backward(loss)
         model.reset_super_network()
